feat(lint-staged): Inline config
Extracts `lint-staged` configuration to a standalone package.
