+++ packages/lint-staged/package.json
@@ -2,7 +2,6 @@
   "name": "@langri-sha/lint-staged",
   "version": "1.0.0-next",
   "private": true,
-  "type": "module",
   "main": "src/index.js",
   "devDependencies": {
     "@langri-sha/tsconfig": "1.0.0-next",
